[Feat] 저장된 인벤토리 아이템 슬롯 불러오기 구현

+++ Assets/01 Scripts/Data/SaveData/PlayerInventorySaveData.cs
@@ -12,7 +12,7 @@ public class PlayerInventorySaveData
     public List<int> ItemIDList;
     public List<int> GunItemAmmoCountList;
     public List<int> QuantityList;
-    public List<float> DurabilityList;
+    public List<int> DurabilityList;
 
     // Money
     public int Money;
