+++ Assets/01 Scripts/Data/SaveData/StorageSaveData.cs
@@ -7,5 +7,5 @@ public class StorageSaveData
     public List<int> ItemIDList;
     public List<int> GunItemAmmoCountList;
     public List<int> QuantityList;
-    public List<float> DurabilityList;
+    public List<int> DurabilityList;
 }
